update to float64

+++ time_test.go
@@ -52,7 +52,7 @@ func TestConvertRoman(t *testing.T) {
 
 func TestSleep(t *testing.T) {
 	Sleep(2)
-	Sleep(float32(0.5))
+	Sleep(0.5)
 }
 
 func TestStrftime(t *testing.T) {
